[US_PA] Deduplicate location metadata (Recidiviz/recidiviz-data#25923)
## Description of the change

https://github.com/Recidiviz/recidiviz-data/pull/25861 fixed the problem
of locations not existing in the location metadata reference view, but
created a new problem where officers with particular role subtypes did
not have hydrated districts, causing the
[current_supervision_staff_missing_district](https://recidiviz-staging.ue.r.appspot.com/admin/validation_metadata/status/details/current_supervision_staff_missing_district?stateCode=US_PA)
validation failure to spike. This is because I incorporated
`locations_from_supervision_contacts` without excluding locations that
already exist in the `supervision_location_ids` reference table, so
there were duplicate entries in `location_metadata` for many locations,
and some did not have hydrated district names.

This change deduplicates the reference view, so the better-hydrated and
less-manual `supervision_location_ids` reference table is the default
source of location metadata, and `locations_from_supervision_contacts`
is only used when there is no other information for a particular
location.

I tested this change in a sandbox to ensure that both the
`current_supervision_staff_missing_district` and the
`location_metadata_missing_locations` pass.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#25918

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1a71830e28926621ced30dc3beeb1d624e239ac2