[Justice Counts] Update README with DB migration instructions (Recidiviz/recidiviz-data#26005)
## Description of the change

Adding instructions to our README about how to generate and apply DB
migrations. One thing in particular of note here is that I added a
command to undo a local migration. We sometimes get into a weird
situation where we run a migration locally during development (which
happens automatically when you run `pipenv run docker-jc`), but then
realize we don't want to commit that migration or we want to make some
changes to it. If you delete the autogenerated migration file, you'll
get an error the next time you run `pipenv run docker-jc` that says
"Cannot locate revision XXX". Instead, what you should do is first run
the command in this PR to undo the migration, and then delete the file.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 6c5941657895fa74117ac67bc3aaf9162c161689