[Bugfix] Properly skip raw data pruning datasets in empty dataset cleanup (Recidiviz/recidiviz-data#26017)
## Description of the change

We have a process that runs every night which deletes any dataset which
is completely empty (allows us to clean up datasets that held sandbox
views/tables which has expired). We had code that was attempting to
exempt that raw data pruning datasets, but was missing a `continue`
statement. As a result, an endpoint that gets called during the flashing
process which expects this dataset to exist will crash.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Fixes
[this](https://console.cloud.google.com/logs/query;cursorTimestamp=2023-12-07T00:02:24.032579763Z;duration=P7D;query=%22%2Fdelete_tables_in_pruning_datasets%22%0Atimestamp%3D%222023-12-06T16:40:06.961427Z%22%0AinsertId%3D%2223dug15tlskh%22?project=recidiviz-123)
exception

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 5beb268edbbcd7252b0ba6a00f519a5b9f9843aa