[Justice Counts] Do not return Missing Sheet warnings when uploading CSV files. (Recidiviz/recidiviz-data#25921)
## Description of the change

Suppress missing sheet warnings when uploading CSV files.

Missing Sheet warnings don't make sense when users upload CSV files
(since only 1 CSV file can be uploaded at a time). This change requires
stringing the `upload_filetype` variable all the way to
`_add_missing_sheet_warnings()`. Note that `upload_filetype` is left as
`None` for most tests when we are not testing functionality that relies
on whether a file was uploaded as a CSV or XLSX.

Task context: https://github.com/Recidiviz/recidiviz-data/issues/20671

Manual Test Results (ran locally):

- [x] "Missing Breakdown/Total Sheet" Warnings are still triggered when
uploading XLSX sheets.
- [x] "Missing Breakdown Sheet" Warnings are suppressed when uploading
CSV of aggregate metrics.
- [x] "Missing Total Sheet" Warnings are suppressed when uploading CSV
of breakdown metrics.

Note to reviewers:
Another possible fix is implemented in this separate PR:
https://github.com/Recidiviz/recidiviz-data/pull/25898
This fix is "cleaner" in that it only requires one line of code. But is
less correct because it also suppresses warnings for single-sheet XLSX
files (which we probably should have warnings for).
Linking it here for discussion.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 74b761e9c6c30a503ee02abaf85910a531f42c93