[US_PA] Update PA SFTP root to use an absolute path (Recidiviz/recidiviz-data#26097)
## Description of the change

The `stat()` command which is called inside of SFTP download code is
expecting an absolute path, not a relative path which we are currently
building because the root path defined in the US_PA delegate is a
relative path.

This PR merely updates the PA delegate to explicitly define the path we
should be searching for files in. A follow-up change could allow us to
handle relative paths more gracefully or enforce that all delegates
define absolute paths for the root search path.

I have confirmed via a test DAG in Airflow that this fixes the PA crash
we were seeing.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Likely fix for Recidiviz/recidiviz-data#26023

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 63594f1a6bd634ba11f282f10b33da3de966d689