[US_ND] Include non-person staff entities in state staff (Recidiviz/recidiviz-data#26111)
## Description of the change

There are entries in `docstars_officers` like "Bismarck CS" and "Grand
Forks CS" that are not individual staff members, so I had initially
excluded them from `state_staff` entity hydration. This caused
normalization to fail because these non-person entities are, as it turns
out, occasionally referenced in the
`state_supervision_period.supervising_officer_staff_id` field. It is
unclear what this means in practice.

This change makes it so that these non-person actors are included in the
`state_staff` entity to avoid errors in normalization, but not in any
other `state_staff` entities (location, supervisor periods, etc.)
because they are not legitimate staff members.

I tested that the normalization pipelines will run successfully after
this change by:
- running a sandbox ingest pipeline for the `docstars_staff` view to
generate a state database that includes this change
(`elise_staff_us_nd_state_primary`)
- using the state dataset generated from the sandbox ingest pipeline to
run normalization in a sandbox calculation pipeline, which ran
successfully (`elise_us_nd_normalized_state`)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 922509825ab776731318bdb58d3add1df7ed7013