Add WARRANT_STATUS enum to supervision type (Recidiviz/recidiviz-data#26121)
## Description of the change

This PR adds the value `WARRANT_STATUS` as an enum value for
`StateSupervisionPeriodSupervisionType`. It also adjusts the description
of `BENCH_WARRANT` as described in the schema change request
[here](https://paper.dropbox.com/doc/APPROVED-Schema-change-request-WARRANT_STATUS-supervision-type-enum--CFedbzjiU35pAoW8UZUPXWXqAg-eZGnswl4sRDQuB4hdrun3).

I made a few assumptions about where to put `WARRANT_STATUS` in a
prioritized ranking of all supervision types which I will call out in
comments below.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#25957

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 41b6e68d64edb0869a92ed10128e16608dd696d6