[US_MI] Fix investigation periods bug (Recidiviz/recidiviz-data#26058)
## Description of the change

This PR changes the `supervision_type` mapping to only map to
`INVESTIGATION` if legal order is NULL, the most recent observed
movement was "Intake", _AND_ supervision level is NULL. Otherwise, if
legal order is NULL and the most recent observed movement was "Intake"
but supervision level is not NULL, map supervision type based on
supervision level.

This PR also includes an upgrade database migration to implement this
change. The migration is kinda lengthy so I'm happy to chat about it if
that makes it easier to review!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 5f619bba9a5a7a56e2ac01350bb8b1e3df62fac7