[Justice Counts] Raise errors  in admin panel when user / agency that already exists (Recidiviz/recidiviz-data#26167)
## Description of the change

Adds two new errors that will be surfaced when a user attempts to create
a new user with an email of an existing user or when they attempt to
create a new agency with the same name of an existing agency.

I no longer am able to download my loom videos, but
[here](https://www.loom.com/share/511baca5fbda44efb6ace9ec781106f0?sid=dad59a3e-68ef-4cc7-aca5-367638451b5f)
is a link of what this looks like end to end!

## Related issues

Closes Recidiviz/recidiviz-data#25701

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0d5edd31ef1a6ead957ffe0cb80766c9bc4857c8