[SCHEMA] Adding DEPORTED enum to supervision_type (Recidiviz/recidiviz-data#26158)
## Description of the change

This PR adds the value DEPORTED as an enum value for
StateSupervisionPeriodSupervisionType.

Based on inspo from Recidiviz/recidiviz-data#26121 , I made a few assumptions about where to put
DEPORTED in a prioritized ranking of all supervision types, which should
be reviewed for agreed upon accuracy.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#25950

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [X] This pull request has a descriptive title and information useful
to a reviewer
- [X] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 435ec641efea3c157652e1a2a70cd4d1dcd90317